pyproject.toml
[tool.poetry]
name = "signtinel"
version = "0.1.0"
description = ""
authors = ["albertocj1 <albertochristianjoshua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
numpy = "^2.3.0"
requests = "^2.32.4"
pandas = "^2.3.0"
scikit-learn = "^1.7.0"
matplotlib = "^3.10.3"
streamlit = "^1.45.1"
seaborn = "^0.13.2"
optuna = "^4.3.0"
hyperopt = "^0.2.7"
ipykernel = "^6.29.5"
chardet = "^5.2.0"
fuzzywuzzy = "^0.18.0"
torch = "^2.7.1"
torchvision = "^0.22.1"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
